+++ front/HW/FEweek07_hw/src/App.jsx
@@ -1,13 +1,19 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import TodoPage from "./pages/TodoPage";
 import TodoDetailPage from "./pages/TodoDetailPage";
+import { useTodos } from "./hooks/useTodos";
 
 function App() {
+  const todoData = useTodos();
+
   return (
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={<TodoPage />} />
-        <Route path="/detail/:todoId" element={<TodoDetailPage />} />
+        <Route path="/" element={<TodoPage {...todoData} />} />
+        <Route
+          path="/detail/:todoId"
+          element={<TodoDetailPage {...todoData} />}
+        />
       </Routes>
     </BrowserRouter>
   );
